Moved Spektrum VTX control stuff to separate source files.
